nullable id

+++ src/main/java/com/cash/examen/controler/UserController.java
@@ -29,7 +29,6 @@ public class UserController {
     private static final String ID_CAN_NOT_BE_NULL = "Id can not be null";
     private static final String PLEASE_SPECIFY_USER_ID = "Please specify the user id";
     private static final String USER_NOT_EXIST = "User id not exist";
-    private static final String ID_IS_A_NUMBER = "User id is a number, please insert the correct value";
     private static final String USER_DELETED = "User has been deleted";
 
     @Autowired
@@ -65,7 +64,7 @@ public class UserController {
         final User user = userService.findUser(id);
         if (!ObjectUtils.allNotNull(user)) {
             log.info("User with id {} not exist", id);
-            return new ResponseEntity(DefaultResponseDTO.builder().status(HttpStatus.CONFLICT).message(ID_IS_A_NUMBER).build(), HttpStatus.CONFLICT);
+            return new ResponseEntity(DefaultResponseDTO.builder().status(HttpStatus.CONFLICT).message(USER_NOT_EXIST).build(), HttpStatus.CONFLICT);
         }
 
         try {
@@ -85,24 +84,32 @@ public class UserController {
             return new ResponseEntity(DefaultResponseDTO.builder().status(HttpStatus.BAD_REQUEST).message(ID_CAN_NOT_BE_NULL).build(), HttpStatus.BAD_REQUEST);
         }
 
+        final User user = userService.findUser(id);
+        if (!ObjectUtils.allNotNull(user)) {
+            log.info("User with id {} not exist", id);
+            return new ResponseEntity(DefaultResponseDTO.builder().status(HttpStatus.CONFLICT).message(USER_NOT_EXIST).build(), HttpStatus.CONFLICT);
+        }
+
         try {
             log.info("Trying to delete User info for id {}", id);
             userService.deleteUser(id);
             return new ResponseEntity(DefaultResponseDTO.builder().status(HttpStatus.OK).message(USER_DELETED).build(), HttpStatus.OK);
         } catch (Exception e) {
+            log.error("Error",e);
             return new ResponseEntity(DefaultResponseDTO.builder().status(HttpStatus.INTERNAL_SERVER_ERROR).message(SERVER_CONFLICT).build(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @RequestMapping(value = ("/create-user"), method = RequestMethod.POST, consumes = "application/json")
+    @RequestMapping(value = ("/users"), method = RequestMethod.POST, consumes = "application/json")
     @ResponseBody
     public ResponseEntity createUser(@RequestBody User user) {
-
+        log.info("Trying to create user email {}", user.getEmail());
         try {
             userService.createUser(user);
+            log.info("Create user with email {}", user.getEmail());
             return new ResponseEntity<>(DefaultResponseDTO.builder().status(HttpStatus.OK).message(USER_CREATED).build(), HttpStatus.CREATED);
         } catch (UserAlreadyRegisteredException e) {
-            log.error("Error", e);
+            log.error("User {} already exist",user.getEmail() , e);
             return new ResponseEntity<>(DefaultResponseDTO.builder().status(HttpStatus.CONFLICT).message(e.getMessage()).build(), HttpStatus.CONFLICT);
         } catch (Exception e) {
             log.error("Error", e);
